+++ src/Camera.h
@@ -26,24 +26,10 @@ public:
 	
 	//method functions
 	
-	Vect getCameraPosition () {return cam_pos;}
-	Vect getCameraDirection () {return cam_dir;}
-	Vect getCameraRight () {return cam_right;}
-	Vect getCameraDown () {return cam_down;}
+	Vect getCameraPosition ();
+	Vect getCameraDirection ();
+	Vect getCameraRight ();
+	Vect getCameraDown ();
 	
 };
-Camera::Camera () {
-	cam_pos  = Vect(0,0,0);
-	cam_dir  = Vect(0,0,1);  // straight out of the screen
-	cam_right= Vect(0,0,0); // specified later
-	cam_down = Vect(0,0,0);  // specified later
-}
-
-Camera::Camera (Vect pos, Vect dir, Vect right, Vect down) {
-	cam_pos  = pos;
-	cam_dir  = dir;
-	cam_right= right;
-	cam_down = down;
-}
-
 #endif /* Camera_h */
