+++ src/Makefile
@@ -1,6 +1,6 @@
 CC = g++
 CFLAGS = -I.
-OBJS = main.o Vect.o Color.o Light.o 
+OBJS = main.o Vect.o Ray.o Color.o Light.o Object.o Plane.o Sphere.o Camera.o
 
 all: raytracer
 
@@ -11,14 +11,29 @@ raytracer: $(OBJS)
 main.o: main.cpp
 	$(CC) -c main.cpp $(CFLAGS)
 
+Ray.o: Ray.cpp
+	$(CC) -c Ray.cpp $(CFLAGS)
+
+Object.o: Object.cpp
+	$(CC) -c Object.cpp $(CFLAGS)
+
 Vect.o: Vect.cpp
 	$(CC) -c Vect.cpp $(CFLAGS)
+
+Plane.o: Plane.cpp
+	$(CC) -c Plane.cpp $(CFLAGS)
+
+Sphere.o: Sphere.cpp
+	$(CC) -c Sphere.cpp $(CFLAGS)
 
 Color.o: Color.cpp
 	$(CC) -c Color.cpp $(CFLAGS)
 
 Light.o: Light.cpp
 	$(CC) -c Light.cpp $(CFLAGS)
+
+Camera.o: Camera.cpp
+	$(CC) -c Camera.cpp $(CFLAGS)
 
 clean:
 	rm -f raytracer $(OBJS)
