+++ src/Object.cpp
@@ -1,0 +1,18 @@
+//
+//  Object.h
+//  Created by Ojashvi Rautela on 11/22/18.
+//
+
+#include "Object.h"
+
+
+Object::Object(){
+}
+
+Color Object::getColor () {
+	return Color (0.0, 0.0, 0.0, 0.0);
+}
+
+double Object::findIntersection(Ray ray){
+	return 0;
+}
