+++ src/Object.h
@@ -12,26 +12,15 @@
 
 class Object{
 	
-	
-	
 	public:
 	
-	Object(){};
+	Object();
 	
 	virtual Color getColor ();
 	
 	//intersect b/w a ray and the object
 	virtual double findIntersection (Ray);
 };
-
-Color Object::getColor () {
-	return Color (0.0, 0.0, 0.0, 0.0);
-}
-
-double Object::findIntersection(Ray ray){
-	return 0;
-}
-
 
 
 #endif /* Vect_h */
